// frontend/amplify/backend/types/amplify-dependent-resources-ref.d.ts
export type AmplifyDependentResourcesAttributes = {
  "api": {
    "courseflexibility": {
      "GraphQLAPIEndpointOutput": "string",
      "GraphQLAPIIdOutput": "string"
    }
  },
  "auth": {
    "courseflexibilityd81a2987": {
      "AppClientID": "string",
      "AppClientIDWeb": "string",
      "IdentityPoolId": "string",
      "IdentityPoolName": "string",
      "UserPoolArn": "string",
      "UserPoolId": "string",
      "UserPoolName": "string"
    }
  },
  "storage": {
    "courseFlexibilityStorage": {
      "BucketName": "string",
      "Region": "string"
    }
  }
}